split up yup file
