+++ back/src/companies/validation.ts
@@ -1,5 +1,5 @@
 import * as yup from "yup";
-import { validDatetime } from "../common/yup";
+import validDatetime from "../common/yup/validDatetime";
 
 export const receiptSchema = yup.object().shape({
   validityLimit: validDatetime({
